requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==19.3.0             # via jsonschema
babel==2.8.0              # via oslo.i18n
certifi==2019.11.28       # via requests, sentry-sdk
chardet==3.0.4            # via requests
coreapi==2.3.3            # via django-rest-swagger, drf-yasg, openapi-codec
coreschema==0.0.4         # via coreapi, drf-yasg
datapunt-config-loader==1.1.2  # via datapunt-objectstore
datapunt-objectstore==2019.4.8  # via -r requirements.in
debtcollector==2.0.0      # via oslo.config, oslo.serialization, oslo.utils, python-keystoneclient
defusedxml==0.6.0         # via djangorestframework-xml
django-choices==1.7.1     # via -r requirements.in
django-extensions==2.2.8  # via -r requirements.in
django-filter==2.2.0      # via -r requirements.in
django-rest-swagger==2.2.0  # via -r requirements.in
django==2.2.11            # via -r requirements.in, django-choices, django-filter, djangorestframework, drf-yasg
djangorestframework-csv==2.1.0  # via drf-amsterdam
djangorestframework-gis==0.15  # via -r requirements.in
djangorestframework-xml==1.4.0  # via drf-amsterdam
djangorestframework==3.11.0  # via -r requirements.in, django-rest-swagger, djangorestframework-csv, djangorestframework-gis, drf-amsterdam, drf-extensions, drf-yasg
drf-amsterdam==0.3.2      # via -r requirements.in
drf-extensions==0.6.0     # via -r requirements.in, drf-amsterdam
drf-yasg==1.17.1          # via -r requirements.in
idna==2.9                 # via requests
inflection==0.3.1         # via drf-yasg
iso8601==0.1.12           # via keystoneauth1, oslo.utils
itypes==1.1.0             # via coreapi
jinja2==2.11.1            # via coreschema
jsonschema==3.2.0         # via datapunt-config-loader
keystoneauth1==3.18.0     # via python-keystoneclient
markupsafe==1.1.1         # via jinja2
msgpack==1.0.0            # via oslo.serialization
netaddr==0.7.19           # via oslo.config, oslo.utils
netifaces==0.10.9         # via oslo.utils
openapi-codec==1.3.2      # via django-rest-swagger
os-service-types==1.7.0   # via keystoneauth1
oslo.config==8.0.1        # via python-keystoneclient
oslo.i18n==4.0.0          # via oslo.config, oslo.utils, python-keystoneclient
oslo.serialization==3.1.0  # via python-keystoneclient
oslo.utils==4.1.0         # via oslo.serialization, python-keystoneclient
packaging==20.3           # via drf-yasg
pbr==5.4.4                # via debtcollector, keystoneauth1, os-service-types, oslo.i18n, oslo.serialization, oslo.utils, python-keystoneclient, stevedore
psycopg2-binary==2.8.4    # via -r requirements.in
pyparsing==2.4.6          # via oslo.utils, packaging
pyrsistent==0.15.7        # via jsonschema
python-keystoneclient==3.22.0  # via -r requirements.in, datapunt-objectstore
python-swiftclient==3.9.0  # via -r requirements.in, datapunt-objectstore
pytz==2019.3              # via babel, django, oslo.serialization, oslo.utils
pyyaml==5.3.1             # via -r requirements.in, datapunt-config-loader, oslo.config, oslo.serialization
requests==2.23.0          # via coreapi, keystoneauth1, oslo.config, python-keystoneclient, python-swiftclient
rfc3986==1.3.2            # via oslo.config
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.10      # via drf-yasg
sentry-sdk==0.14.3        # via -r requirements.in
simplejson==3.17.0        # via django-rest-swagger
six==1.14.0               # via debtcollector, django-choices, django-extensions, djangorestframework-csv, drf-yasg, jsonschema, keystoneauth1, oslo.config, oslo.i18n, oslo.serialization, oslo.utils, packaging, python-keystoneclient, python-swiftclient, stevedore
sqlparse==0.3.1           # via django
stevedore==1.32.0         # via keystoneauth1, oslo.config, python-keystoneclient
unicodecsv==0.14.1        # via djangorestframework-csv
uritemplate==3.0.1        # via coreapi, drf-yasg
urllib3==1.25.8           # via requests, sentry-sdk
wrapt==1.12.1             # via debtcollector
xlrd==1.2.0               # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
